pyproject.toml
[project]
name = "code-breaker-ai-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.125.0",
    "langchain>=1.1.2",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "yapf>=0.43.0",
]

[dependency-groups]
test = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]
